% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_plant_available.R
\name{background_soil_N_modifier}
\alias{background_soil_N_modifier}
\title{background_soil_N_modifier}
\usage{
background_soil_N_modifier(N_back_before, N_back, t)
}
\arguments{
\item{N_back_before}{N background before (other year) or 1.4*N_yield}

\item{N_back}{calculated N back based on residues and mineralization}

\item{t}{timestep}
}
\value{
list with soil N humus and N losses
}
\description{
modifies N background to accept input N back (N_back_before)
if N_back_before only has 1 length, it automatically assumes 0 N loss and converts it to list (as input_background_soil_N)
otherwise it accepts N_back_before input (length=2)
if timestep != 1 just receives calculated N back
}
\note{
NOT USED
}
