% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grass_Below_ground_production.R
\name{belowground_yield_DM}
\alias{belowground_yield_DM}
\title{belowground_yield_DM}
\usage{
belowground_yield_DM(yield_dm, f_harvested = 0.8, f_shoot)
}
\arguments{
\item{yield_dm}{DM yield of harvested aboveground material (Ygs ;kg ha-1 yr-1)}

\item{f_harvested}{fraction of aboveground production harvested (??; \%) (0-1)}

\item{f_shoot}{fraction of total DM production that is aboveground (??gDM; shoot fraction) (0-1)}
}
\description{
computes belowground production (Ygr)
}
