% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization_helpers.R
\name{calculate_max_yield_DM}
\alias{calculate_max_yield_DM}
\title{calculate_max_yield_DM}
\usage{
calculate_max_yield_DM(df_obs)
}
\arguments{
\item{df_obs}{dataframe of site-specific observations}
}
\value{
maximum yield DM
}
\description{
computes maximum Yield DM based on a quadratic function
}
\note{
Done by Jeroen
}
