% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clover_N_fixation.R
\name{clover_harvested_N_fixation}
\alias{clover_harvested_N_fixation}
\title{clover_harvested_N_fixation}
\usage{
clover_harvested_N_fixation(clover_N_yield, N_min_clover, f_Navail = 0.5)
}
\arguments{
\item{clover_N_yield}{N in aboveground clover returned to soil (kg N ha-1 yr-1)}

\item{N_min_clover}{unclear what this is; soil N to clover? (kg N ha-1 yr-1)}

\item{f_Navail}{Proportion of available N  +  fixed N recovered in clover yield (alpha_c) (0-1)}
}
\description{
Computes N fixed by clover
}
