% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clover_N_residues.R
\name{clover_N_mineralization}
\alias{clover_N_mineralization}
\title{clover_N_mineralization}
\usage{
clover_N_mineralization(N_availability, clover_N_total, grass_N_uptake)
}
\arguments{
\item{N_availability}{N availability; see plant_N_avaible() (kg N ha-1 yr-1)}

\item{clover_N_total}{total crop N of clover (kg N ha-1 yr-1)}

\item{grass_N_uptake}{total grass N uptake (kg N ha-1 yr-1)}
}
\description{
computes N mineralization from clover
}
