% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clover_Above_ground_production.R
\name{clover_N_total}
\alias{clover_N_total}
\title{clover_N_total}
\usage{
clover_N_total(N_yield, f_ab_clover)
}
\arguments{
\item{N_yield}{clover N yield (;kg N ha-1 yr-1)}

\item{f_ab_clover}{N yield/totl crop N for clover (\%)}
}
\description{
computes total N of clover
}
