% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clover_Above_ground_production.R
\name{clover_N_yield_base}
\alias{clover_N_yield_base}
\title{clover_N_yield}
\usage{
clover_N_yield_base(yield_clover_dm_max = 6000, f_clover, f_n_clover = 0.04)
}
\arguments{
\item{f_clover}{portion of clover occupation  (0-1) (pc)}

\item{yield_clover_max}{maximum clover yield in monoculture (Yc ;kg ha-1 yr-1)}

\item{f_n}{the concentration of N in the harvested material (0-1) (nc; kg kg DM-1 yr-1)}
}
\value{
clover_N_yield_base(6000, 0.3, 0.04)
}
\description{
computes clover N yield (accounts for portion of clover in land)
}
