% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clover_Above_ground_production.R
\name{clover_N_yield_soil_aboveground}
\alias{clover_N_yield_soil_aboveground}
\title{clover_N_yield_soil_aboveground}
\usage{
clover_N_yield_soil_aboveground(N_yield, f_Navail = 0.5)
}
\arguments{
\item{N_yield}{clover N yield (;kg N ha-1 yr-1)}

\item{f_Navail}{Proportion of available N  +  fixed N recovered in clover yield (alpha_c) (0-1)}
}
\description{
computes N in aboveground clover returned to soil
}
