% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clover_N_fixation.R
\name{clover_Ndfa}
\alias{clover_Ndfa}
\title{clover_Ndfa}
\usage{
clover_Ndfa(clover_N, total_N_fixation)
}
\arguments{
\item{clover_N}{crop N of clover (kg N ha-1 yr-1)}

\item{total_N_fixation}{total N fixed by clover (kg N ha-1 yr-1); NOT clover_harvested_N_fixation()}
}
\description{
Computes N2 fixed per aboveground N biomass
}
