% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_plant_available.R
\name{excess_soil_N}
\alias{excess_soil_N}
\title{excess_soil_N}
\usage{
excess_soil_N(N_availability, grass_Nuptake, clover_N_min)
}
\arguments{
\item{N_availability}{grass N availability; see plant_N_avaible() (kg N ha-1 yr-1)}

\item{grass_Nuptake}{total grass N uptake (kg N ha-1 yr-1)}

\item{clover_N_min}{clover N mineralization (kg N ha-1 yr-1)}
}
\description{
mass balance to compute excess N in soil
}
