% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clover_Above_ground_production.R
\name{fraction_clover}
\alias{fraction_clover}
\title{fraction_clover}
\usage{
fraction_clover(grass_yield_dm, grass_yield_dm_max, over_yielding_factor = 0.5)
}
\arguments{
\item{grass_yield_dm}{grass DM yield of harvested aboveground material (Ygs ;kg ha-1 yr-1)}

\item{grass_yield_dm_max}{grass max value of yield_dm when N supply is not a limiting factor (Yg,max; kg ha-1 yr-1)}

\item{f_displacement}{#Displacement rate of clover by grass (0-1) (epsilon)}
}
\description{
computes portion of clover occupation
}
