% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization_helpers.R
\name{generate_random_params}
\alias{generate_random_params}
\title{generate_random_params}
\usage{
generate_random_params(n = 10000)
}
\arguments{
\item{n}{length of params}
}
\value{
dataframe with min_t, min_t1 and min back parametrization
}
\description{
generates normal dist of residue fractions that always sum to 1
}
