% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grass_Above_ground_production.R
\name{grass_N_uptake}
\alias{grass_N_uptake}
\title{total_plant_N_uptake}
\usage{
grass_N_uptake(f_shoot, N_yield)
}
\arguments{
\item{f_shoot}{fraction of total DM production that is aboveground (??gDM; shoot fraction) (0-1)}

\item{N_yield}{N in harvestable yield (kg N ha-1 yr-1)}

\item{N_residues_t}{N residues input at year t (kg N ha-1 yr-1)}
}
\description{
computes total N plant uptake based on the residues and harvested N yield
}
