% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_functions.R
\name{grass_only_simulation}
\alias{grass_only_simulation}
\title{grass_only_simulation}
\usage{
grass_only_simulation(
  yield_dm_max = 12000,
  N_fert = 100,
  f_min_t = 0.6,
  f_min_t1 = 0.25,
  N_back_before = NULL,
  N_min_t1 = NULL
)
}
\arguments{
\item{yield_dm_max}{max yield dm; max threshold}

\item{N_fert}{N fertilisation (the higher, the more the model reaches equilibrium)}

\item{f_min_t}{fraction of crop residues mineralized at timestep t}

\item{f_min_t1}{fraction of crop residues mineralized at timestep t+1 (next year)}

\item{N_back_before}{N background; set this up only in the first year, otherwise it is from previous year}
}
\description{
loop until model reaches equilibrium where plant available N is similar to total crop N uptake
}
