% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grass_parametrization.R
\name{grass_simulation_metrics}
\alias{grass_simulation_metrics}
\title{grass_simulation_metrics}
\usage{
grass_simulation_metrics(
  N_fert = 100,
  year = 1,
  f_min_t,
  f_min_t1,
  df_obs,
  df_obs_Nback,
  yield_dm_max,
  N_min_t1 = NULL
)
}
\arguments{
\item{N_fert}{N fertilisation level (eg, 100)}

\item{year}{int (0,n)}

\item{f_min_t}{fraction of mineralized at timestep t, sigma}

\item{f_min_t1}{fraction of mineralized at timestep t+1, beta g}

\item{df_obs}{S23 obs for a given site and fertilisation level}

\item{df_obs_Nback}{S23 obs for a given site}

\item{yield_dm_max}{maximum yield DM (kg ha-1 yr-1)}

\item{N_min_t1}{N mineralized in the year before}
}
\value{
list with different output; see function output
}
\description{
function that runs simulation of grass only for a given f_min_t and f_min_t1 and returns sum of nRMSE
of simulated and observed data
}
