% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grass_parametrization.R
\name{grassclover_simulation_metrics}
\alias{grassclover_simulation_metrics}
\title{Title}
\usage{
grassclover_simulation_metrics(
  species = c("blanca", "s100"),
  N_fert = 100,
  year = 1,
  grass_f_min_t,
  clover_f_min_t,
  f_min_t1,
  df_obs,
  df_obs_Nback,
  grass_yield_dm_max = 12000,
  N_min_t1 = NULL
)
}
\arguments{
\item{species}{}

\item{N_fert}{}

\item{year}{}

\item{grass_f_min_t}{}

\item{clover_f_min_t}{}

\item{f_min_t1}{}

\item{df_obs}{}

\item{df_obs_Nback}{}

\item{grass_yield_dm_max}{}

\item{N_min_t1}{}
}
\description{
Title
}
