% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_plant_available.R
\name{input_background_soil_N}
\alias{input_background_soil_N}
\title{input_background_soil_N}
\usage{
input_background_soil_N(ef_loss, N_residues_t, N_min_t, N_min_t1)
}
\arguments{
\item{ef_loss}{\% leaching+denitrification losses (0-1) (??)}

\item{N_residues_t}{N residues input at year t (kg N ha-1 yr-1)}

\item{N_min_t}{mineralization from residues during the same growing period at year t (kg N ha-1 yr-1); see input_mineralization_same_year()}

\item{N_min_t1}{N mineralization from residues that will occur next yea (kg N ha-1 yr-1); see input_mineralization_following_year()}
}
\value{
list with soil N humus and N losses
}
\description{
soil N supply from residues are the remaining N from residues after (i) mineralization at t, (ii) mineralization that will occur at t+1 and (iii) N losses
}
\note{
Not used
}
