% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_plant_available.R
\name{input_mineralization_following_year}
\alias{input_mineralization_following_year}
\title{input_mineralization_following_year}
\usage{
input_mineralization_following_year(ef_loss, f_min_t, f_min_t1, N_residues_t)
}
\arguments{
\item{ef_loss}{\% leaching+denitrification losses (0-1) (??)}

\item{f_min_t}{fraction of residues mineralized in the same growing period (0-1) (??g)}

\item{f_min_t1}{fraction of residues mineralized in the next year (0-1) (??g)}

\item{N_residues_t}{N residues input at year t (kg N ha-1 yr-1)}

\item{t}{timestep; if timestep = 1, then there is no previousl Nmin}
}
\value{
list with min for the following year and N losses
}
\description{
computes N mineralization from residues that will occur next year (t+1 or t1)
Includes (i) N losses through denitrification/leaching, (ii) the fraction of residues that will be mineralized at t+1
(iii) N mineralization from residues at t
}
