% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_plant_available.R
\name{input_mineralization_same_year}
\alias{input_mineralization_same_year}
\title{input_mineralization_same_year}
\usage{
input_mineralization_same_year(f_min_t, N_residues_t)
}
\arguments{
\item{f_min_t}{fraction of residues mineralized in the same growing period (0-1) (??g)}

\item{N_residues_t}{N residues input at year t (kg N ha-1 yr-1)}
}
\description{
computes N mineralization from residues during the same growing period
}
\note{
initial min from previous residues set to 2 like jeroen
}
