% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_plant_available.R
\name{input_Nfertiliser}
\alias{input_Nfertiliser}
\title{input_Nfertiliser}
\usage{
input_Nfertiliser(N_fert, ef_ammonia)
}
\arguments{
\item{N_fert}{N fertiliser input (kg N ha-1 yr-1)}

\item{ef_ammonia}{\% N-NH3 losses (0-1) (??)}
}
\value{
listh with effective N input and N-NH3 losses
}
\description{
computes effecive N from fertilisers
excludes N-NH3 emissions
}
