% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grass_Above_ground_production.R
\name{N_conc_harvested_fitted}
\alias{N_conc_harvested_fitted}
\title{N_conc_harvested_fitted}
\usage{
N_conc_harvested_fitted(
  yield_dm,
  yield_dm_max,
  f_n_min = 0.01903,
  lambda = 0.01138
)
}
\arguments{
\item{lambda}{quadratic parameter}
}
\description{
computes N concentrastion in harvested material using a quadratic function fitted to the relative yield (yield dm / yield dm max)
}
\details{
@param yield_dm DM yield of harvested aboveground material (Ygs ;kg ha-1 yr-1)
@param yield_dm_max max value of yield_dm when N supply is not a limiting factor (Yg,max; kg ha-1 yr-1)
@param f_n_min min value of the N concentration in harvested material where is not harvestabled growth (0-1) (ngs,min)
}
