% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grass_Below_ground_production.R
\name{N_concentration_roots_fitted}
\alias{N_concentration_roots_fitted}
\title{N_concentration_roots_fitted}
\source{
fitted data used data from Schenk and the Gorissen
}
\usage{
N_concentration_roots_fitted(k_gr = 0.2413266, f_n, f_gr_min = 0.2257085)
}
\arguments{
\item{k_gr}{coefficient}

\item{f_n}{the concentration of N in the harvested material (0-1) (ngs; kg kg DM-1 yr-1)}

\item{f_gr_min}{Concentration of N in the root material when the N in shoot is extrapolated to zero (ngrmin, kg N kg DM-1)}
}
\description{
fits N concentration in roots based on fitted data rom Schenk and the Gorissen
}
