% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Humus.R
\name{N_humus_next_year}
\alias{N_humus_next_year}
\title{N_humus_next_year}
\usage{
N_humus_next_year(
  f_min_t,
  yield_dm,
  yield_dm_belowground,
  f_harvested = 0.8,
  N_conversion = 0.006141
)
}
\arguments{
\item{f_min_t}{fraction of residues mineralized in the same year}

\item{yield_dm}{aboveground yield dm}

\item{yield_dm_belowground}{belowground yield dm}

\item{f_harvested}{fraction of aboveground production harvested (??; \%) (0-1)}

\item{N_conversion}{C:N conversion flows from C-TOOL}
}
\description{
computes N bound in humus from the decomposition of residues at timestep t+1
}
