% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grass_Below_ground_production.R
\name{N_input_roots}
\alias{N_input_roots}
\title{N_input_roots}
\usage{
N_input_roots(yield_dm, f_harvested = 0.8, f_shoot, f_n_roots)
}
\arguments{
\item{yield_dm}{DM yield of harvested aboveground material (Ygs ;kg ha-1 yr-1)}

\item{f_harvested}{fraction of aboveground production harvested (??; \%) (0-1)}

\item{f_shoot}{fraction of total DM production that is aboveground (??gDM; shoot fraction) (0-1)}

\item{f_n_roots}{N concentration in roots (ng; kg N kg DM-1)}
}
\description{
computes the N input from roots (Ngr)
}
