% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_residues.R
\name{N_residues_availability_next_year}
\alias{N_residues_availability_next_year}
\title{N_residues_availability_next_year}
\usage{
N_residues_availability_next_year(f_min_t, N_res_total, N_humus_t1)
}
\arguments{
\item{f_min_t}{fraction of residues mineralized in the same year; gamma; (0-1)}

\item{N_res_total}{total N from residues (aboveground + roots)}

\item{N_humus_t1}{(N bound in humus in the current year)}
}
\description{
computes mineral N available for plant uptake from the decomposition of residues in the following year
}
