% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_residues.R
\name{N_residues_in}
\alias{N_residues_in}
\title{N_residues_in}
\usage{
N_residues_in(N_in_aboveground, N_in_roots)
}
\arguments{
\item{N_in_aboveground}{N into the soil from aboveground residues; see N_yield_soil_aboveground()}

\item{N_in_roots}{N into the soil from belowground residues; see N_input_roots()}
}
\description{
computes N into the soil from above- and belowground residues (N_residuesg)
}
\note{
See Above_ground_production.R and below_ground_production.R
}
