% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grass_Above_ground_production.R
\name{N_yield}
\alias{N_yield}
\title{N_yield}
\usage{
N_yield(yield_dm, f_n)
}
\description{
computes N uptake in harvested material
}
\details{
@param yield_dm DM yield of harvested aboveground material (Ygs ;kg ha-1 yr-1)
@param f_n the concentration of N in the harvested material (0-1) (ngs; kg kg DM-1 yr-1)
}
