% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grass_Above_ground_production.R
\name{N_yield_combined}
\alias{N_yield_combined}
\title{N_yield_combined}
\usage{
N_yield_combined(yield_dm, yield_dm_max, f_n_min, f_n_max)
}
\description{
@param yield_dm DM yield of harvested aboveground material (Ygs ;kg ha-1 yr-1)
@param yield_dm_max max value of yield_dm when N supply is not a limiting factor (Yg,max; kg ha-1 yr-1)
@param f_n_min min value of the N concentration in harvested material where is not harvestabled growth (0-1) (ngs,min)
@param f_n_max max value of the n concentration in harvested material where N is saturated (0-1) (ngs,max)
@description
combines N_yield() and N_conc_harvested()
@unit kg N ha-1 yr-1
}
