% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grass_Above_ground_production.R
\name{N_yield_soil_aboveground}
\alias{N_yield_soil_aboveground}
\title{N_yield_soil_aboveground}
\usage{
N_yield_soil_aboveground(N_yield, f_harvested = 0.8)
}
\arguments{
\item{N_yield}{N yield of harvested material (Ngs; kg N ha-1 yr-1)}

\item{f_harvested}{fraction of aboveground production harvested (??; \%) (0-1)}
}
\description{
Computes N entering the soil aboveground; excludes fraction of aboveground production harvested
}
\note{
f_harvested is defaulted to 80\% (mechanical harvesting; this would be lower for grazing)
}
