% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_plant_available.R
\name{plant_N_avaible_NOTUSED}
\alias{plant_N_avaible_NOTUSED}
\title{plant_N_avaible}
\usage{
plant_N_avaible_NOTUSED(N_eff_fert, N_min_t, N_min_t1, N_back)
}
\arguments{
\item{N_eff_fert}{N effective fertiliser after NH3 volatilization (kg N ha-1 yr-1); see input_Nfertiliser}

\item{N_min_t}{N mineralisation from residues at t (kg N ha-1 yr-1); see input_mineralization_same_year}

\item{N_min_t1}{N mineralisation from residues from t-1 (kg N ha-1 yr-1); see input_mineralization_following_year}

\item{N_back}{N available from semi-stable SON (kg N ha-1 yr-1); see input_background_soil_N}
}
\description{
calculates plant available N
}
