% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grass_Below_ground_production.R
\name{shoot_fraction}
\alias{shoot_fraction}
\title{shoot_fraction}
\usage{
shoot_fraction(yield_dm, total_yield_dm, f_harvested = 0.8)
}
\arguments{
\item{yield_dm}{DM yield of harvested aboveground material (Ygs ;kg ha-1 yr-1)}

\item{total_yield_dm}{total yield (sum of above- and below ground material); see total_DM_production()}

\item{f_harvested}{fraction of aboveground production harvested (??; \%) (0-1)}
}
\description{
cmputes fraction of the total DM production that is aboveground (??gDM; shoot fraction)

cmputes fraction of the total DM production that is aboveground (??gDM; shoot fraction)
}
