% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grass_Below_ground_production.R
\name{shoot_fraction_fitted}
\alias{shoot_fraction_fitted}
\title{shoot_fraction_fitted}
\usage{
shoot_fraction_fitted(k_agdm = 0.09792063, f_n, f_shoot_dm_min = 0.2853004)
}
\arguments{
\item{k_agdm}{coefficient
@param f_n the concentration of N in the harvested material (0-1) (ngs; kg kg DM-1 yr-1)
@param f_shoot_dm_min shoot fraction when N concentration is extrapolated to 0}
}
\description{
computes fitted shoot fraction using data from (Schenk et al., 1996) and (Gorissen and Cotrufo, 1999)
k_agdm and f_shoot_dm_min are the fitted coefficients
}
