% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grass_Below_ground_production.R
\name{total_DM_production}
\alias{total_DM_production}
\title{total_DM_production}
\usage{
total_DM_production(yield_dm, f_harvested = 0.8, yield_below)
}
\arguments{
\item{yield_dm}{DM yield of harvested aboveground material (Ygs ;kg ha-1 yr-1)}

\item{f_harvested}{fraction of aboveground production harvested (??; \%) (0-1)}

\item{yield_below}{DM yield of new root material and exudates (Ygr; kg ha-1 yr-1)}
}
\description{
computes total DM production, ie, sum of aboveground and below ground production (Ygtot)
}
